Added support for a module Auth cookie.
Once the user is successfully validated against the NLM service, we
cache that knowledge and issue the user a cookie.  Future requests from
the user, during the cookies livetime will allow the user access without
us having to query the NLM service again.  Should improve performance a
little, and reduce our vulnerability for short outages in the NLM
service availability.